Move contiguous_iterator.hpp to container/detail
and expose range_snapshotting_terator as typdef in containers.
